Remove Custom Footnotes and Custom Annotations
This is a temporary change so that the package can be stable and publishable. Custom footnotes and annotations will be major features once added, and so should be given more attention (but not hold back the first release of this package).

This doesn't remove the ability to create footnotes within the `message` in LilyPond.