Initial HTML export
This exports a basic html file, with the option to specify which properties are written. Annotations and their properties are wrapped in div tags with classes specified automatically, leaving ID open for the moment. The point of id for individual annotations will obviously be to link to score items once that is eventually approached. Different annotation lists (as/in bookparts) in a book could also use id's to differentiate themselves from one another.