Reduce redundant code in "div-open"
This approach includes the necessary conditionals in the
bindings of the parameters themselves, avoiding to have
to separate but nearly identical structures in the branches of the
if clause.

Additionally it fixes an issue when both class and id are present
(where the space was missing in between). Now an arbitrary number of tags can be joined in an opening tag