Make \with {} block in annotations optional
If no with block is present a bare one is created with just the ann-type
attribute.

This is of no use for actual scholarly work but often it makes sense
to add such empty annotations as a quick hack to point to something.
And: I intend to provide the option of default content (e.g. for 'author),
that would lend itself to that approach.